+++ CHANGELOG.md
@@ -24,3 +24,5 @@
 - `0.3.2` Add user IDs, --control-server for managed use and --opt-out to disable tracking.
 - `0.3.2` Add user IDs, --control-server for managed use and --opt-out to disable tracking.
 - `0.3.3` Fix `inspect` mode, bug fixes.
+- `0.3.4` Various fixes. Scanning of built-in IDE tools.
+- `0.3.5` Improving description of some built-in IDE tools.